Implement EventTraceScheduledDispatcherService for improved event tracing and memory monitoring
