+++ projects/oneteme/inspect-ng-collector/src/lib/log.service.ts
@@ -1,25 +1,28 @@
-// log.service.ts
-import { Injectable } from '@angular/core';
-import { SessionManager } from './session-manager.service';
+import {Injectable} from '@angular/core';
 import {createLogEntry} from './util';
+import {EventTraceScheduledDispatcherService} from "./event-trace-scheduled-dispatcher.service";
+import {SessionManager} from "./session-manager.service";
 
 @Injectable({ providedIn: 'root' })
 export class LogService {
-  constructor(private readonly sessionManager: SessionManager) {}
+
+  constructor(private readonly dispatcher: EventTraceScheduledDispatcherService,
+              private readonly sessionManager: SessionManager) {
+  }
 
   info(message: string) {
     if (message) {
-      this.sessionManager.traceQueue.push(createLogEntry("INFO", message));
+      this.dispatcher.addToQueue(createLogEntry("INFO", message, this.sessionManager.currentSessionID()));
     }
   }
   warn(message: string) {
     if (message) {
-      this.sessionManager.traceQueue.push(createLogEntry("WARN", message));
+      this.dispatcher.addToQueue(createLogEntry("WARN", message, this.sessionManager.currentSessionID()));
     }
   }
   error(message: string) {
     if (message) {
-      this.sessionManager.traceQueue.push(createLogEntry("ERROR", message));
+      this.dispatcher.addToQueue(createLogEntry("ERROR", message, this.sessionManager.currentSessionID()));
     }
   }
 }
